% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{create_index}
\alias{create_index}
\title{Create file index}
\usage{
create_index(path, pattern = NULL, all.files = FALSE, recursive = FALSE,
  verbose = TRUE)
}
\arguments{
\item{path}{Directory to index.  This can be be the current
directory (".") or any other.}

\item{pattern}{An optional pattern (passed to \code{\link{dir}})}

\item{all.files}{An optional logical indicating of "dot files"
should be included (passed to \code{\link{dir}})}

\item{recursive}{An optional logical indicating that the directory
listing should be recursive.}

\item{verbose}{Be verbose when computing hashes (which can take a
while for very large directories).}
}
\value{
This is called for the side effect of creating an index
  file (\code{.remotefile/index.json}) within the directory
  \code{path}.
}
\description{
Create file index
}

