% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{is_remotefile_dest}
\alias{is_remotefile_dest}
\title{Check if a path is a remotefile destination}
\usage{
is_remotefile_dest(path)
}
\arguments{
\item{path}{Path to test (a directory)}
}
\description{
Check if a path is a remotefile destination
}

