% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{remotefile_fetch}
\alias{remotefile_fetch}
\title{Fetch a remote file}
\usage{
remotefile_fetch(filename, obj)
}
\arguments{
\item{filename}{The name of a file, to be found at the remote
location \code{obj}.  Do not include the first part of the URL.}

\item{obj}{A \code{remotefile} object, as created by
\code{\link{remotefile_init}}}
}
\value{
The filename, on your computer
}
\description{
Fetches remote files.
}

