% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{remotefile_init}
\alias{remotefile_init}
\title{Initialise local remotefile repo}
\usage{
remotefile_init(path, src, refresh = FALSE)
}
\arguments{
\item{path}{Path to locally store files}

\item{src}{The URL for the remote}

\item{refresh}{Force a refresh of the index?}
}
\description{
Initialise a local remotefile repo
}

