% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{remotefile_list}
\alias{remotefile_list}
\title{List remote files}
\usage{
remotefile_list(obj, pattern = NULL, local = FALSE)
}
\arguments{
\item{obj}{A \code{remotefile} object, as created by
\code{\link{remotefile_init}}}

\item{pattern}{Optional pattern, as a regular expression.  Will be
interpreted by \code{grep} or \code{dir}.}

\item{local}{Look in the local directory?}
}
\description{
List remote files
}

